refactor(knockdog): features/map -> features/kindergarten-map 도메인에 종속적이게 디렉터리 변경
- features/kindergarten-map - 유치원 메인 지도 관련 기능 레이어
- shared/ui/map - 지도 관련 공통 ui
