+++ apps/knockdog/src/features/kindergarten-map/model/useMapUrlState.ts
@@ -13,12 +13,20 @@ const CENTER_PARSER = createParser<{ lat: number; lng: number }>({
   eq: (a, b) => a.lat === b.lat && a.lng === b.lng,
 });
 
-export function useMapState() {
+/**
+ * Kindergarten Map URL 상태를 관리하는 훅
+ *
+ * @description
+ * - center: 지도 중심 좌표
+ * - zoomLevel: 지도 줌레벨
+ * - searchedLevel: 이전 검색에 사용된 행정구역 스케일
+ */
+export function useMapUrlState() {
   const [center, setCenter] = useQueryState('center', CENTER_PARSER);
   const [zoomLevel, setZoomLevel] = useQueryState('zoom', parseAsInteger);
 
   /**
-   * 이전 검색에 사용된 행정구역 단위 스케일
+   * 이전 검색에 사용된 행정구역 스케일
    * 1: 광역시/도, 2: 시/군/구, 3: 읍/면/동
    *
    * 현재 지도 줌레벨과 비교하여 행정구역 단위가 변경되었을 때만
