+++ apps/knockdog/src/features/kindergarten-map/ui/CurrentLocationDisplayFAB.tsx
