--- apps/knockdog/src/features/map/ui/CurrentSelectionMarker.tsx
@@ -1,10 +1,0 @@
-export function CurrentSelectionMarker() {
-  return (
-    <svg width='36' height='47' viewBox='0 0 36 47' fill='none' xmlns='http://www.w3.org/2000/svg'>
-      <circle cx='18' cy='18' r='18' fill='#FF6E0C' fillOpacity='0.3' />
-      <circle cx='18' cy='18' r='13' fill='#FF6E0C' stroke='#FF6E0C' />
-      <circle cx='18' cy='18' r='3.5' fill='white' stroke='#FF6E0C' />
-      <path d='M18 31C18 36.5646 18 47.5508 18 46.9785' stroke='#FF6E0C' strokeWidth='3' />
-    </svg>
-  );
-}
