--- apps/knockdog/src/views/kindergarten-main-page/config/map.ts
@@ -1,5 +1,0 @@
-// TODO: 실제 기본 좌표로 변경 필요!!
-export const DEFAULT_MAP_CENTER = { lat: 37.54, lng: 127.07 } as const;
-
-// TODO: 실제 기본 줌 레벨로 변경 필요!!
-export const DEFAULT_MAP_ZOOM_LEVEL = 15;
